Inline `newtype::name!` macro to its single caller
